LayoutTests/inspector/replay/window-navigator-plugins-memoized-expected.txt

Tests that we can capture and replay nondeterminism in window.navigator.plugins and window.navigator.mimeTypes.

 Waiting for test page to finish its initial load...
Test page initial load done.
Waiting for capturing to start...
Capturing has started.
Waiting to capture initial navigation...
Initial navigation captured. Dumping state....
Capture stopped, now starting replay to completion...
Playback has started.
Waiting to replay initial navigation...
Initial navigation replayed. Dumping state...
PASS: Nondeterministic state should not differ during capture and replay.
